dom/tests/mochitest/localstorage/test_localStorageQuotaPrivateBrowsing.html

<html xmlns="http://www.w3.org/1999/xhtml">
<head>
<title>localStorage and DOM quota test</title>

<script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
<script type="text/javascript" src="pbSwitch.js"></script>
<script type="text/javascript" src="interOriginTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />

<script type="text/javascript">

netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");

var currentTest = 1;
var prefs = Components.classes["@mozilla.org/preferences-service;1"]
            .getService(Components.interfaces.nsIPrefBranch);
var quota;
try {
  quota = prefs.getIntPref("dom.storage.default_quota");
} catch (ex) {
  quota = 5*1024;
}
prefs.setIntPref("dom.storage.default_quota", 1);

enterPrivateBrowsing();

function doNextTest()
{
  if (!get_PBSvc()) {
    // There is no PB service, skip all tests
    currentTest = 999;
    ok(true, "No private browsing service, test could not be performed");
  }

  slave = frame;

  netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");

  switch (currentTest)
  {
    // Initialy setup the quota to testing value of 1024B and
    // set a 500 bytes key with name length 1 (allocate 501 bytes)
    case 1:
      slaveOrigin = "http://example.com";
      slave.location = slaveOrigin + slavePath + "frameQuota.html?add&A&success";
      break;

    // In subdomain now set another key with length 500 bytes, i.e.
    // allocate 501 bytes
    case 2:
      slaveOrigin = "http://test1.example.com";
      slave.location = slaveOrigin + slavePath + "frameQuota.html?add&B&success";
      break;

    // Try to set the same key value again to check we don't fail
    // even 1002 bytes has already been exhausted from the quota
    // We just change the value of an existing key.
    case 3:
      slaveOrigin = "http://test1.example.com";
      slave.location = slaveOrigin + slavePath + "frameQuota.html?add&B&success";
      break;

    // Try to set the same key to a larger value that would lead to
    // quota reach and check that the value is still the old one
    case 4:
      slaveOrigin = "http://test1.example.com";
      slave.location = slaveOrigin + slavePath + "frameQuota.html?add2&B&failure";
      break;

    // In a different subdomain try to set a new 500 bytes key
    // and check we fail because we are over the quota
    case 5:
      slaveOrigin = "https://test2.example.com";
      slave.location = slaveOrigin + slavePath + "frameQuota.html?add&C&failure";
      break;

    // Remove from the second subdomain the second key, it must not fail
    // This should release the allocated space of the quota assigned to
    // example.com.
    case 6:
      slaveOrigin = "http://test1.example.com";
      slave.location = slaveOrigin + slavePath + "frameQuota.html?remove&B&success";
      break;

    // Now try again to set 500 bytes key, it must succeed.
    case 7:
      slaveOrigin = "https://test2.example.com";
      slave.location = slaveOrigin + slavePath + "frameQuota.html?add&C&success";
      break;

    case 8:
      // Do a clean up...
      // TODO Bug 455070, use just ?clear what invokes call
      // of clear() in the target frame. W/o clear method we must
      // call clear implemented as removeItem for each item in
      // the localStorage.
      slaveOrigin = "http://example.com";
      slave.location = slaveOrigin + slavePath + "frameQuota.html?clear&A&";
      break;

    case 9:
      // Do a clean up...
      slaveOrigin = "http://test1.example.com";
      slave.location = slaveOrigin + slavePath + "frameQuota.html?clear&B&";
      break;

    case 10:
      // Do a clean up...
      slaveOrigin = "https://test2.example.com";
      slave.location = slaveOrigin + slavePath + "frameQuota.html?clear&C&";
      break;

    default:
      prefs.setIntPref("dom.storage.default_quota", quota);

      leavePrivateBrowsing();

      SimpleTest.finish();
  }

  ++currentTest;
}

function doStep()
{
}

SimpleTest.waitForExplicitFinish();

</script>

</head>

<body onload="doNextTest();">
  <iframe src="" name="frame"></iframe>
</body>
</html>
